# src/openapi_doc.json
{
  "info": {
    "description": "\r\n  ",
    "version": "1.0.0",
    "title": "SL_aiohttp Swagger Doc"
  },
  "tags": [
    {
      "name": "Index",
      "description": ""
    }
  ],
  "paths": {
    "/": {
      "get": {
        "tags": [
          "Index"
        ],
        "summary": "Show index p",
        "responses": {
          "200": {
            "description": "OK"}
        },
        "parameters": []
      }
    },
    "/login": {
      "post": {
        "tags": [
          "User auth"
        ],
        "summary": "Create a new user",
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "OK"
          },
          "400": {
            "description": "Failed."
          }
        },
        "parameters": [
          {
            "description": "User Request Post Object",
            "required": true,
            "name": "body",
            "in": "body",
            "schema": {
              "properties": {
                "login": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "consumes": [
          "application/json"
        ]
      }
    },
    "/logout": {
      "get": {
        "tags": [
          "User auth"
        ],
        "summary": "Logout",
        "responses": {
          "200": {
            "description": "OK",
            "schema": 0
          }
        },
        "parameters": []
      }
    },
    "/users": {
      "get": {
        "tags": [
          "User Request"
        ],
        "summary": "Returns all available users in the system",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/users list"
            }
          }
        },
        "parameters": []
      },
      "post": {
        "tags": [
          "User Request"
        ],
        "summary": "Create a new user",
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "OK"
          },
          "400": {
            "description": "Failed."
          }
        },
        "parameters": [
          {
            "description": "User Request Post Object",
            "required": true,
            "name": "body",
            "in": "body",
            "schema": {"$ref": "#/definitions/user"
            }
          }
        ],
        "consumes": [
          "application/json"
        ]
      }
    },
    "/users/{login}": {
      "get": {
        "tags": [
          "User Request"
        ],
        "summary": "Get user by login",
        "parameters": [
          {
            "in": "path",
            "name": "login",
            "required": true,
            "description": "User Request login"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/user"
            }
          },
          "400": {
            "description": "Failed."
          },
          "404": {
            "description": "Not found."
          }
        }
      },
      "post": {
        "summary": "Update user by login",
        "tags": [
          "User Request"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "login",
            "required": true,
            "description": "user login"
          },
          {
            "description": "User Request Object",
            "required": true,
            "name": "body",
            "in": "body",
            "schema": {"$ref": "#/definitions/user"


            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Failed.."
          }
        },
        "consumes": [
          "application/json"
        ]
      },
      "delete": {
        "summary": "Delete user by login",
        "tags": [
          "User Request"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "login",
            "required": true,
            "description": "user Request login"
          }
        ],
        "responses": {
          "204": {
            "description": "OK"
          },
          "400": {
            "description": "Failed."
          },
          "404": {
            "description": "Not found."
          }
        }
      }
    }
  },
  "swagger": "2.0",
  "definitions": {
    "login_form": {
      "properties": {
        "login": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "login": {
      "properties": {
        "login": {
          "type": "string"
        }
      }
    },
    "user": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "format": "string"
        },
        "surname": {
          "type": "string",
          "format": "string"
        },
        "login": {
          "type": "string",
          "format": "string"
        },
        "password": {
          "type": "string",
          "format": "string"
        },
        "birthdate": {
          "type": "string",
          "format": "date"
        },
        "blocked": {
          "type": "boolean"
        },
        "admin": {
          "type": "boolean"
        },
        "readonly": {
          "type": "boolean"
        }
      }
    },
    "users list": {
      "type": "object",
      "properties": {
        "user list": {
          "type": "object",
          "additionalProperties": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "format": "string"
              },
              "surname": {
                "type": "string",
                "format": "string"
              },
              "login": {
                "type": "string",
                "format": "string"
              },
              "password": {
                "type": "string",
                "format": "string"
              },
              "birthdate": {
                "type": "string",
                "format": "date"
              },
              "blocked": {
                "type": "boolean"
              },
              "admin": {
                "type": "boolean"
              },
              "readonly": {
                "type": "boolean"
              }
            }
          }
        }
      }
    }
  },
  "x-components": {}
}